requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
deprecation==2.1.0
distlib==0.3.9
Django==5.2.3
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.2
gotrue==2.12.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jalali_core==1.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jdatetime==5.2.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
mysql-connector-python==9.3.0
packaging==25.0
pbs-installer==2025.6.10
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
postgrest==1.0.2
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
realtime==2.4.3
requests==2.32.4
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
websockets==14.2
Werkzeug==3.0.1
yarl==1.20.0
zstandard==0.23.0
# کتابخانه‌های امنیتی (اختیاری - برای Google OAuth)
# bcrypt==4.1.3
# Flask-Limiter==3.8.0
# google-auth==2.30.0
# google-auth-oauthlib==1.2.1
# google-auth-httplib2==0.2.0
